+++ azbfs/zt_url_directory_test.go
@@ -126,7 +126,7 @@ func (dus *DirectoryUrlSuite) TestCreateDirectoryAndFiles(c *chk.C) {
 
 	// Create fileUrl from directoryUrl and create file inside the directory
 	fileUrl, _ := getFileURLFromDirectory(c, dirUrl)
-	fresp, err := fileUrl.Create(context.Background())
+	fresp, err := fileUrl.Create(context.Background(), azbfs.BlobFSHTTPHeaders{})
 	defer delFile(c, fileUrl)
 
 	c.Assert(err, chk.IsNil)
@@ -175,7 +175,7 @@ func (dus *DirectoryUrlSuite) TestDirectoryStructure(c *chk.C) {
 
 	// Create a file inside directory
 	fileUrl, _ := getFileURLFromDirectory(c, dirUrl)
-	fresp, err := fileUrl.Create(context.Background())
+	fresp, err := fileUrl.Create(context.Background(), azbfs.BlobFSHTTPHeaders{})
 	defer delFile(c, fileUrl)
 
 	c.Assert(err, chk.IsNil)
@@ -188,7 +188,7 @@ func (dus *DirectoryUrlSuite) TestDirectoryStructure(c *chk.C) {
 
 	// create a file inside the sub-dir created above
 	subDirfileUrl, _ := getFileURLFromDirectory(c, subDirUrl)
-	fresp, err = subDirfileUrl.Create(context.Background())
+	fresp, err = subDirfileUrl.Create(context.Background(), azbfs.BlobFSHTTPHeaders{})
 	defer delFile(c, subDirfileUrl)
 
 	c.Assert(err, chk.IsNil)
@@ -230,7 +230,7 @@ func (dus *DirectoryUrlSuite) TestListDirectoryWithSpaces(c *chk.C) {
 
 	// Create a file inside directory
 	fileUrl, _ := getFileURLFromDirectory(c, dirUrl)
-	_, err = fileUrl.Create(context.Background())
+	_, err = fileUrl.Create(context.Background(), azbfs.BlobFSHTTPHeaders{})
 	defer delFile(c, fileUrl)
 
 	// list the directory created above.